refactor(routing): move bandit_router from cost/ → routing/ (audit-193)
Moves `bandit_router.py` (1256 LOC) from `core/cost/` to `core/routing/`.
`cost/` now holds only spend tracking; routing decisions live under
`routing/`. Updates the back-compat shim in `_REDIRECT_MAP` and the
three non-shim call sites (cost.py x2 and test_cost_engine.py x2) to
the new import path. All other callers use the
`bernstein.core.bandit_router` shim form which continues to work
unchanged.
